Wrappers: a way to specify possible output for traffic
Wrappers are a way to specify how some generated traffic should be
output, either via a given protocolar stack or a socket or sent to some
sink or what not.

Those wrappers are easy to serialize (both so that it can be specified
by the user and saved in a file).

Still TBD: all but socket wrappers :)